Vinicius.company

+++ OnTheFlyApp.CompanyService/Controllers/CompaniesServiceController.cs
@@ -4,6 +4,7 @@ using System.Net;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Bson;
+using MongoDB.Driver;
 using OnTheFly.Models;
 using OnTheFly.Models.Dto;
 using OnTheFlyApp.CompanyService.Service;
@@ -23,7 +24,7 @@ namespace OnTheFlyApp.CompanyService.Controllers
             _util = util;
         }
 
-        [HttpGet("acitevated/", Name = "GetAll")]
+        [HttpGet("getActivated/", Name = "GetAllActivated")]
         public ActionResult<List<CompanyGetDTO>> GetAll()
         {
             List<CompanyGetDTO> lstReturn = _companyService.GetAll();
@@ -32,7 +33,7 @@ namespace OnTheFlyApp.CompanyService.Controllers
             return lstReturn;
         }
 
-        [HttpGet("disable/", Name = "GetDisable")]
+        [HttpGet("getDisabled/", Name = "GetDisable")]
         public ActionResult<List<CompanyGetDTO>> GetDisable()
         {
             List<CompanyGetDTO> lstReturn = _companyService.GetDisable();
@@ -41,19 +42,26 @@ namespace OnTheFlyApp.CompanyService.Controllers
             return lstReturn;
         }
 
-        [HttpGet("company/{cnpj}", Name = "GetCnpj")]
+        [HttpGet("getCompanyCnpj/", Name = "GetCnpj")]
         public ActionResult<CompanyGetDTO> GetCnpj(string cnpj)
         {
             cnpj = _util.JustDigits(cnpj);
+            cnpj = cnpj.Replace("%2", "");
+
+            if (_util.VerifyCnpj(cnpj) == false)
+                return BadRequest("Cnpj inválido");
+
             var companyreturn = _companyService.GetByCompany(cnpj);
-            if(companyreturn == null) { companyreturn = _companyService.GetByCompanyRestricted(cnpj).Result;
+            if (companyreturn == null)
+            {
+                companyreturn = _companyService.GetByCompanyRestricted(cnpj).Result;
                 if (companyreturn != null) { return Ok("RESTRICTED"); }
             }
 
             return companyreturn == null ? NotFound("Companhia não encontrada") : companyreturn;
         }
 
-        [HttpGet("restricted/", Name = "GetRestricted")]
+        [HttpGet("getRestricted/", Name = "GetRestricted")]
         public ActionResult<List<CompanyGetDTO>> GetRestricted()
         {
             List<CompanyGetDTO> lstReturn = _companyService.GetRestricted();
@@ -62,10 +70,11 @@ namespace OnTheFlyApp.CompanyService.Controllers
             return lstReturn;
         }
 
-        [HttpPost]
+        [HttpPost("postCompany/")]
         public ActionResult<CompanyGetDTO> Create(CompanyGetDTO companydto)
         {
             companydto.Cnpj = _util.JustDigits(companydto.Cnpj);
+            companydto.Cnpj = companydto.Cnpj.Replace("%2", "");
 
             if (_util.VerifyCnpj(companydto.Cnpj) == false)
                 return BadRequest("Cnpj inválido");
@@ -78,7 +87,7 @@ namespace OnTheFlyApp.CompanyService.Controllers
             { return BadRequest("Cnpj já está registrado"); }
 
             if (_companyService.GetByCompanyRestricted(companydto.Cnpj).Result != null)
-            { return BadRequest("Cnpj está cadastrado e restrito"); }
+            { return BadRequest("Cnpj está cadastrado e RESTRITO"); }
 
             if (companydto.NameOpt == "string" || companydto.NameOpt == string.Empty) { companydto.NameOpt = companydto.Name; }
 
@@ -102,44 +111,58 @@ namespace OnTheFlyApp.CompanyService.Controllers
             company.Address = new(newAddress);
             CompanyGetDTO companyReturn = new();
 
-            try
-            {
-                companyReturn = new(_companyService.Create(company));
-            }
-            catch (Exception)
-            {
-                return BadRequest("Endereço já cadastrado");
-            }
+            companyReturn = new(_companyService.Create(company));
+
             return companyReturn;
         }
 
-        [HttpPut("cnpj/{cnpj}", Name = "Status")]
+        [HttpPut("patchStatus/", Name = "Status")]
         public async Task<ActionResult<CompanyGetDTO>> Update(string cnpj)
         {
+            cnpj = _util.JustDigits(cnpj);
+            cnpj = cnpj.Replace("%2", "");
+
+            if (_util.VerifyCnpj(cnpj) == false)
+                return BadRequest("Cnpj inválido");
+
             var companyStatus = _companyService.GetByCompany(cnpj);
 
             if (companyStatus == null) { companyStatus = await _companyService.GetByCompanyRestricted(cnpj); }
-            if (companyStatus.Cnpj == null) return NotFound("Companhia não encontrada");
+            if (companyStatus == null) return NotFound("Companhia não encontrada");
 
-            CompanyGetDTO companyReturn = new();
-
-            if (companyStatus.Status == true)
+            try
             {
-                companyReturn = _companyService.Update(cnpj, false).Result;
+                if (companyStatus.Status == true)
+                {
+                    CompanyGetDTO companyReturn = await _companyService.Update(cnpj, false);
+                    if (companyReturn == null) { return BadRequest("Companhia sem avião"); }
+
+                    return Ok(companyReturn);
+                }
+                else
+                {
+                    CompanyGetDTO companyReturn = await _companyService.Update(cnpj, true);
+                    if (companyReturn == null) { return BadRequest("Companhia sem avião");  }
+
+                    return Ok(companyReturn);
+                }
             }
-            else
+            catch (Exception)
             {
-                companyReturn = _companyService.Update(cnpj, true).Result;
+                return Problem(HttpStatusCode.InternalServerError.ToString());
             }
 
-            if (companyReturn == null) { return BadRequest("Companhia não possui avião"); }
-
-            return Ok(companyReturn);
         }
 
-        [HttpPut("restricted/{cnpj}")]
+        [HttpPut("patchRestrictedCnpj/")]
         public async Task<ActionResult> UpdateRestricted(string cnpj)
         {
+            cnpj = _util.JustDigits(cnpj);
+            cnpj = cnpj.Replace("%2", "");
+
+            if (_util.VerifyCnpj(cnpj) == false)
+                return BadRequest("Cnpj inválido");
+
             CompanyGetDTO company = new();
 
             company = _companyService.GetByCompany(cnpj);
@@ -161,18 +184,24 @@ namespace OnTheFlyApp.CompanyService.Controllers
         }
 
         [HttpDelete]
-        public HttpStatusCode Delete(string cnpj)
+        public ActionResult Delete(string cnpj)
         {
+            cnpj = _util.JustDigits(cnpj);
+            cnpj = cnpj.Replace("%2", "");
+
+            if (_util.VerifyCnpj(cnpj) == false)
+                return BadRequest("Cnpj inválido");
+
             var companyResult = _companyService.GetByCompany(cnpj);
             if (companyResult == null)
             {
                 companyResult = _companyService.GetByCompanyRestricted(cnpj).Result;
-                if (companyResult == null) return HttpStatusCode.NotFound;
+                if (companyResult == null) return NotFound("Não localizado");
             }
 
             _companyService.Delete(cnpj);
 
-            return HttpStatusCode.OK;
+            return Ok("Deletado");
         }
     }
 }
