+++ OnTheFlyApp.CompanyService/Service/CompaniesService.cs
@@ -16,7 +16,7 @@ namespace OnTheFlyApp.CompanyService.Service
         private readonly IMongoCollection<Company> _companyDisabled;
         private readonly IMongoCollection<Address> _address;
         static readonly HttpClient companyClient = new HttpClient();
-        static readonly string endpointAirCraft = "https://localhost:5002/api/AirCraftsService/company/";
+        static readonly string endpointAirCraft = "https://localhost:5002/api/AirCraftsService/cnpjCompany?cnpj=";
         public CompaniesService() { }
         public CompaniesService(ICompanyServiceSettings settings)
         {
@@ -57,7 +57,7 @@ namespace OnTheFlyApp.CompanyService.Service
             List<Company> lstRestricted = _companyRestricted.Find(c => true).ToList();
             List<CompanyGetDTO> lstReturn = new();
 
-            foreach(var company in lstRestricted) { lstReturn.Add(companyDTO = new(company)); }
+            foreach (var company in lstRestricted) { lstReturn.Add(companyDTO = new(company)); }
 
             return lstReturn;
         }
@@ -70,7 +70,7 @@ namespace OnTheFlyApp.CompanyService.Service
             var companyDisabled = _companyDisabled.Find<Company>(c => c.Cnpj == cnpj).FirstOrDefault();
             if (companyDisabled == null && company == null) return null;
 
-            if(companyDisabled != null) { return companyReturn = new(companyDisabled); }
+            if (companyDisabled != null) { return companyReturn = new(companyDisabled); }
             else return companyReturn = new(company);
 
         }
@@ -88,7 +88,13 @@ namespace OnTheFlyApp.CompanyService.Service
         public Company Create(Company company)
         {
             if (_address.Find(a => a.Number == company.Address.Number && a.ZipCode == company.Address.ZipCode).FirstOrDefault() != null)
-                throw new Exception();
+            {
+                company.Status = false;
+                _companyDisabled.InsertOne(company);
+
+                return company;
+            }
+
             company.Status = false;
             _address.InsertOne(new Address(company.Address));
             _companyDisabled.InsertOne(company);
@@ -98,54 +104,69 @@ namespace OnTheFlyApp.CompanyService.Service
 
         public async Task<CompanyGetDTO> Update(string cnpj, bool status)
         {
-            var company = _companyRestricted.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
-            if(company != null)
+            if (status)
             {
+                var companyRestricted = await _companyRestricted.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync();
+                var companyDisabled = await _companyDisabled.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync();
+
                 HttpResponseMessage responseAirCraft = await CompaniesService.companyClient.GetAsync(endpointAirCraft + cnpj);
                 if (!responseAirCraft.StatusCode.ToString().Equals("OK"))
-                    return null;
+                { throw new Exception(); }
 
-                CompanyGetDTO companyRestricted = new();
-                if (status == true) { company.Status = true; _companyRestricted.ReplaceOne(c => c.Cnpj == cnpj, company); 
-                    return companyRestricted = new(company); }
-                else company.Status = false; _companyRestricted.ReplaceOne(c => c.Cnpj == cnpj, company); 
-                    return companyRestricted = new(company);
+                if (responseAirCraft.Content.Headers.ContentLength == 2) { return null; }
+
+                if (companyRestricted != null)
+                {
+                    CompanyGetDTO companyActivated = new();
+
+                    _companyRestricted.DeleteOne(c => c.Cnpj == cnpj);
+                    companyRestricted.Status = true; _company.InsertOne(companyRestricted);
+                    return companyActivated = new(companyRestricted);
+                }
+                else
+                {
+                    CompanyGetDTO companyActivated = new();
+
+                    _companyDisabled.DeleteOne(c => c.Cnpj == cnpj);
+                    companyDisabled.Status = true; _company.InsertOne(companyDisabled);
+                    return companyActivated = new(companyDisabled);
+                }
             }
-
-            if (status == true)
+            else
             {
-                HttpResponseMessage responseAirCraft = await CompaniesService.companyClient.GetAsync(endpointAirCraft + cnpj);
-                if (!responseAirCraft.StatusCode.ToString().Equals("204"))
-                    return null;
+                var companyRestricted = _companyRestricted.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
+                var companyActivated = _company.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
 
-                company = _companyDisabled.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
-                company.Status = status;
-                CompanyGetDTO companyTrue = new(company);
-                _companyDisabled.DeleteOne(c => c.Cnpj == cnpj);
-                _company.InsertOne(company);
+                CompanyGetDTO companyDisabled = new();
+                if (companyRestricted != null)
+                {
+                    companyRestricted.Status = status; _companyRestricted.ReplaceOne(c => c.Cnpj == cnpj, companyRestricted);
+                    return companyDisabled = new(companyRestricted);
+                }
 
-                return companyTrue;
-            }
-            else {
-                var insertDisabled = _company.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
-                _company.DeleteOne(c => c.Cnpj == cnpj);
-                insertDisabled.Status = status;
-                await _companyDisabled.InsertOneAsync(insertDisabled);
-
-                CompanyGetDTO companyFalse = new(insertDisabled);
-
-                return companyFalse; 
+                else
+                {
+                    companyActivated.Status = status; _company.DeleteOne(c => c.Cnpj == cnpj);
+                    _companyDisabled.InsertOne(companyActivated);
+                    return companyDisabled = new(companyActivated);
+                }
             }
         }
 
         public async void UpdateRestricted(CompanyGetDTO company, string cnpj, bool status)
         {
             Company companyInsert = new(company);
-            if (company.Status == true && status == true){ _company.DeleteOne(c => c.Cnpj == cnpj); 
-                _companyRestricted.InsertOne(companyInsert); return;}
-            
-            if(company.Status == false && status == true){ _companyDisabled.DeleteOne(c => c.Cnpj == cnpj); 
-                _companyRestricted.InsertOne(companyInsert); return;}
+            if (company.Status == true && status == true)
+            {
+                _company.DeleteOne(c => c.Cnpj == cnpj);
+                _companyRestricted.InsertOne(companyInsert); return;
+            }
+
+            if (company.Status == false && status == true)
+            {
+                _companyDisabled.DeleteOne(c => c.Cnpj == cnpj);
+                _companyRestricted.InsertOne(companyInsert); return;
+            }
 
             CompanyGetDTO companyResult = new(await _companyRestricted.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync());
             if (companyResult == null) { throw new Exception(); }
@@ -155,17 +176,17 @@ namespace OnTheFlyApp.CompanyService.Service
             else _companyRestricted.DeleteOne(c => c.Cnpj == cnpj); _companyDisabled.InsertOne(companyInsert); return;
         }
 
-        public void Delete(string cnpj)
+        public async void Delete(string cnpj)
         {
-            var deletedDisabled = _companyDisabled.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
             var deletedCompany = _company.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
+            var deletedDisabled = _companyDisabled.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
             if (deletedDisabled == null && deletedCompany == null)
             {
                 var deleteRestricted = _companyRestricted.Find(c => c.Cnpj == cnpj).FirstOrDefaultAsync().Result;
-                if (deleteRestricted != null) { _companyRestricted.DeleteOne(cnpj); _companyDeleted.InsertOne(deletedDisabled); return; }
+                if (deleteRestricted != null) { _companyRestricted.DeleteOne(c => c.Cnpj == cnpj); _companyDeleted.InsertOne(deleteRestricted); return; }
             }
-            if(deletedDisabled != null) { _companyDisabled.DeleteOne(cnpj); _companyDeleted.InsertOne(deletedDisabled); return; }
-            else _company.DeleteOne(cnpj); _companyDeleted.InsertOne(deletedCompany); return;
+            if (deletedDisabled != null) { _companyDisabled.DeleteOne(c => c.Cnpj == cnpj); _companyDeleted.InsertOne(deletedDisabled); }
+            else { _company.DeleteOne(c => c.Cnpj == cnpj); _companyDeleted.InsertOne(deletedCompany); }
         }
     }
 }
